🔨내가 만든 펀딩 목록 조회 test까지 구현!(#55)

+++ fundings/models.py
@@ -170,8 +170,8 @@ class Reward(models.Model):
     content = models.CharField(
          max_length=50,
     )
-    amount = models.PositiveSmallIntegerField(
-        choices=RewardAmountChoices.choices,
+    amount = models.PositiveIntegerField(
+    choices=RewardAmountChoices.choices,
     )
     expired_at = models.DateTimeField(
         null=True,
